requirements.txt
python-dotenv
langchain-google-genai
crewai
crewai-tools
google-generativeai
pdfplumber
spacy
langchain-groq